+++ hpc/nvidia/cuda/fakeCUDA/utils.hpp
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: GPL-3.0
+/* Copyright (c) 2025 Rong Tao */
+#include <sys/types.h>
+
+
+void memdump(void *mem, size_t size);