+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 #### Breaking Changes
 #### Added
 - fakeCUDA: add runtime.hpp;
+- fakeCUDA: add utils.cpp;
 #### Changed
 #### Deprecated
 #### Removed